updates to and info for
added the PDF for an example to the info needed on the street light repair form.

changes on the plans and programs page are proposed for web forms . will have to copy and paste them into the live pages. there are other changes that are on the live pages that make it out of sync.

building a services page for 2 of the form that will be added in the future.

the questions page may end up turning into a web form in the future.
side note  testing functionality of a few things with forms and cross page capabilites

the validate . php will have to be done for each form uniqueness doesnt allow for the ease of on validation page.